Refactor GraphicsSet to GraphicsPage and util
